Move tooling from llvm library to its own library
