+++ .gitignore
@@ -4,6 +4,6 @@ check.log
 /bin
 /build
 .Makefile.override
-libjlm/include/jlm/tooling/CommandPaths.hpp
+jlm/tooling/CommandPaths.hpp
 .idea
 docs/html
