+++ Makefile.sub
@@ -52,6 +52,7 @@ $(JLM_BUILD)/%.a:
 include $(JLM_ROOT)/jlm/hls/Makefile.sub
 include $(JLM_ROOT)/jlm/llvm/Makefile.sub
 include $(JLM_ROOT)/jlm/rvsdg/Makefile.sub
+include $(JLM_ROOT)/jlm/tooling/Makefile.sub
 include $(JLM_ROOT)/jlm/util/Makefile.sub
 include $(JLM_ROOT)/jlc/Makefile.sub
 include $(JLM_ROOT)/jlm-opt/Makefile.sub
