+++ jlc/Makefile.sub
@@ -13,8 +13,8 @@ jlc-release: CXXFLAGS += -O3
 jlc-release: $(JLM_BIN)/jlc
 
 $(JLM_BIN)/jlc: CPPFLAGS += -I$(JLM_ROOT) -I$(shell $(LLVMCONFIG) --includedir)
-$(JLM_BIN)/jlc: LDFLAGS += $(shell $(LLVMCONFIG) --libs core irReader) $(shell $(LLVMCONFIG) --ldflags) $(shell $(LLVMCONFIG) --system-libs) -L$(JLM_BUILD)/ -lllvm -lrvsdg -lutil
-$(JLM_BIN)/jlc: $(patsubst %.cpp, $(JLM_BUILD)/%.o, $(JLC_SRC)) $(JLM_BUILD)/librvsdg.a $(JLM_BUILD)/libllvm.a $(JLM_BUILD)/libutil.a
+$(JLM_BIN)/jlc: LDFLAGS += $(shell $(LLVMCONFIG) --libs core irReader) $(shell $(LLVMCONFIG) --ldflags) $(shell $(LLVMCONFIG) --system-libs) -L$(JLM_BUILD)/ -ltooling -lllvm -lrvsdg -lutil
+$(JLM_BIN)/jlc: $(patsubst %.cpp, $(JLM_BUILD)/%.o, $(JLC_SRC)) $(JLM_BUILD)/libtooling.a $(JLM_BUILD)/librvsdg.a $(JLM_BUILD)/libllvm.a $(JLM_BUILD)/libutil.a
 	@mkdir -p $(JLM_BIN)
 	$(CXX) $(CXXFLAGS) $(CPPFLAGS) -o $@ $^ $(LDFLAGS)
