+++ jlm-hls/Makefile.sub
@@ -38,8 +38,8 @@ jlm-hls-release: CXXFLAGS += -O3
 jlm-hls-release: $(JLM_BIN)/jlm-hls
 
 $(JLM_BIN)/jlm-hls: CPPFLAGS += -I$(JLM_ROOT) -I$(shell $(LLVMCONFIG) --includedir)
-$(JLM_BIN)/jlm-hls: LDFLAGS += $(shell $(LLVMCONFIG) --libs core irReader) $(shell $(LLVMCONFIG) --ldflags) $(shell $(LLVMCONFIG) --system-libs) -L$(JLM_BUILD)/ -lhls -lllvm -lrvsdg -lutil
-$(JLM_BIN)/jlm-hls: $(patsubst %.cpp, $(JLM_BUILD)/%.o, $(JLM_HLS_SRC)) $(JLM_BUILD)/librvsdg.a $(JLM_BUILD)/libhls.a $(JLM_BUILD)/libllvm.a $(JLM_BUILD)/libutil.a
+$(JLM_BIN)/jlm-hls: LDFLAGS += $(shell $(LLVMCONFIG) --libs core irReader) $(shell $(LLVMCONFIG) --ldflags) $(shell $(LLVMCONFIG) --system-libs) -L$(JLM_BUILD)/ -ltooling -lhls -lllvm -lrvsdg -lutil
+$(JLM_BIN)/jlm-hls: $(patsubst %.cpp, $(JLM_BUILD)/%.o, $(JLM_HLS_SRC)) $(JLM_BUILD)/libtooling.a $(JLM_BUILD)/librvsdg.a $(JLM_BUILD)/libhls.a $(JLM_BUILD)/libllvm.a $(JLM_BUILD)/libutil.a
 	@mkdir -p $(JLM_BIN)
 	$(CXX) $(CXXFLAGS) $(CPPFLAGS) -o $@ $^ $(LDFLAGS)
